feat(pack.resiliation): rename pack-resiliation fileto index
