+++ packages/manager/apps/telecom/src/app/telecom/pack/resiliation/index.js
